pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "raspbot"
version = "0.1.0"
description = "Clean Python library for controlling the Yahboom Raspbot V2 robot car"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Raspbot contributors" }]
keywords = ["raspberry-pi", "robot", "yahboom", "raspbot", "hardware", "i2c"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
dependencies = [
    "smbus2>=0.4",
]

[project.optional-dependencies]
oled = [
    "luma.oled>=3.8",
    "Pillow>=9.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.urls]
Repository = "https://github.com/your-org/raspbot"
Issues = "https://github.com/your-org/raspbot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/raspbot"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["src/raspbot"]
omit = ["tests/*"]
